refactor(live2d): split live2d player action into multiple files
