+++ src/utils/Live2DPlayer/action/special_effect/StopShakeWindow.ts
@@ -1,0 +1,18 @@
+import type { Live2DController } from "../../Live2DController";
+import type { Snippet } from "../../../../types.d";
+import { log } from "../../log";
+
+export default async function StopShakeWindow(
+  controller: Live2DController,
+  action: Snippet
+) {
+  const action_detail =
+    controller.scenarioData.SpecialEffectData[action.ReferenceIndex];
+  log.log(
+    "Live2DController",
+    "SpecialEffect/StopShakeWindow",
+    action,
+    action_detail
+  );
+  controller.layers.dialog.stop_shake();
+}
