+++ src/utils/Live2DPlayer/action/special_effect/Telop.ts
@@ -1,0 +1,14 @@
+import type { Live2DController } from "../../Live2DController";
+import type { Snippet } from "../../../../types.d";
+import { log } from "../../log";
+
+export default async function Telop(
+  controller: Live2DController,
+  action: Snippet
+) {
+  const action_detail =
+    controller.scenarioData.SpecialEffectData[action.ReferenceIndex];
+  log.log("Live2DController", "SpecialEffect/Telop", action, action_detail);
+  controller.layers.telop.draw(action_detail.StringVal);
+  await controller.layers.telop.show(300, true);
+}
